refactor(api): add invitation routes to core Api.ts typed contract
- Add InvitationResult and InvitationSummaryRow interfaces to Api.ts
- Register all 6 invitation routes in APIGet and APIPost
- Ground invitationThunks.ts casts by re-exporting InvitationResult from Api.ts
- Ground invitationsListThunks.ts casts by re-exporting InvitationSummaryRow from Api.ts
- Derive RedemptionLookupResult and RedemptionAcceptResult in redeemInvitationThunks.ts
  directly from the APIGet/APIPost index types (no duplicate interface definitions)
